move code to pvlib modules, rename data file
