--- pvlib/data/inverter_fit_snl_datasheet.csv
@@ -1,19 +1,0 @@
-,fraction_of_rated_power,efficiency,dc_voltage_level,pdc,pac
-0,0.1,0.8921460663497615,Vmin,112.08926853105183,100.0
-1,0.1,0.8764140086437499,Vnom,114.10132541668285,100.0
-2,0.1,0.861227163890917,Vmax,116.11338354472292,100.0
-3,0.2,0.9252558010464728,Vmin,216.1564399529278,200.0
-4,0.2,0.91667390617836,Vnom,218.18009507198235,200.0
-5,0.2,0.9082497361753336,Vmax,220.20375237564716,200.0
-6,0.3,0.9369099565845942,Vmin,320.20152832361623,300.0
-7,0.3,0.930993739729357,Vnom,322.23632361610834,300.0
-8,0.3,0.9251517631120498,Vmax,324.2711217355865,300.0
-9,0.5,0.9465654129565353,Vmin,528.225512105162,500.0
-10,0.5,0.9428959295527041,Vnom,530.2812159102147,500.0
-11,0.5,0.9392547809878062,Vmax,532.3369229743545,500.0
-12,0.75,0.9516178182868241,Vmin,788.1315225372807,750.0
-13,0.75,0.9491138280981478,Vnom,790.2108027472997,750.0
-14,0.75,0.946622992013311,Vmax,792.2900735855503,750.0
-15,1.0,0.9542895293424474,Vmin,1047.9000023075275,1000.0
-16,1.0,0.9523809523809523,Vnom,1050.0,1000.0
-17,1.0,0.9504799924028748,Vmax,1052.0999999925673,1000.0
