+++ pvlib/data/inverter_fit_snl_sim.csv
@@ -1,0 +1,19 @@
+fraction_of_rated_power,efficiency,dc_voltage_level,dc_voltage,dc_power,ac_power,efficiency
+0.1,0.892146066,Vmin,220,112.0892685,100,0.892146067
+0.1,0.876414009,Vnom,240,114.1013254,100,0.876414009
+0.1,0.861227164,Vmax,260,116.1133835,100,0.861227164
+0.2,0.925255801,Vmin,220,216.15644,200,0.925255801
+0.2,0.916673906,Vnom,240,218.1800951,200,0.916673906
+0.2,0.908249736,Vmax,260,220.2037524,200,0.908249736
+0.3,0.936909957,Vmin,220,320.2015283,300,0.936909957
+0.3,0.93099374,Vnom,240,322.2363236,300,0.93099374
+0.3,0.925151763,Vmax,260,324.2711217,300,0.925151763
+0.5,0.946565413,Vmin,220,528.2255121,500,0.946565413
+0.5,0.94289593,Vnom,240,530.2812159,500,0.94289593
+0.5,0.939254781,Vmax,260,532.336923,500,0.939254781
+0.75,0.951617818,Vmin,220,788.1315225,750,0.951617818
+0.75,0.949113828,Vnom,240,790.2108027,750,0.949113828
+0.75,0.946622992,Vmax,260,792.2900736,750,0.946622992
+1,0.954289529,Vmin,220,1047.900002,1000,0.95428953
+1,0.952380952,Vnom,240,1050,1000,0.952380952
+1,0.950479992,Vmax,260,1052.1,1000,0.950479992
